backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-gis>=1.0
django-filter>=23.5
psycopg2-binary>=2.9.9
Pillow>=10.1.0
networkx>=3.2.1
